Solution Architect AI Agent/requirements.txt
# CORE SERVER
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0

# GOOGLE CLOUD
langchain-google-vertexai==1.0.5
google-cloud-aiplatform==1.55.0

# LANGCHAIN STABLE STACK
langchain==0.2.5
langchain-core==0.2.9
langchain-community==0.2.5
langgraph==0.1.5

# DATA & UTILS
chromadb==0.5.0
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
aiofiles==23.2.1